requirements.txt
asgiref==3.5.2
Django==3.2
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
django-logentry-admin==1.1.0
django-markdownx==3.0.1
Markdown==3.3.7
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
